Revert "Merge pull request #156 from bahrain-bp/advance-view"
This reverts commit 7297c065d7145c3ec34217951ee76c7f4f4ec362, reversing
changes made to 4ee547288db31a833d6365b36c2a09039a622780.
