+++ package.json
@@ -41,21 +41,20 @@
     "@aws-sdk/client-bedrock-runtime": "^3.706.0",
     "@aws-sdk/client-cloudformation": "^3.709.0",
     "@aws-sdk/client-comprehend": "^3.687.0",
-    "@aws-sdk/client-dynamodb": "^3.705.0",
-    "@aws-sdk/client-lex-runtime-v2": "^3.693.0",
+    "@aws-sdk/client-dynamodb": "^3.705.0",    
     "@aws-sdk/client-sns": "^3.714.0",
+    "@aws-sdk/client-lex-runtime-v2": "^3.693.0",
     "@aws-sdk/client-textract": "^3.687.0",
+    "aws-cdk": "^2.172.0",
     "@aws-sdk/credential-provider-node": "^3.699.0",
     "@aws-sdk/lib-dynamodb": "^3.705.0",
     "@aws-sdk/types": "^3.696.0",
-    "aws-cdk": "^2.172.0",
     "crawler-request": "^1.2.2",
     "csv-parser": "^3.0.0",
     "date-fns": "^4.1.0",
     "express": "^4.21.1",
     "html2canvas": "^1.4.1",
     "jspdf": "^2.5.2",
-    "lucide-react": "^0.469.0",
     "mime-types": "^2.1.35",
     "pdf-lib": "^1.17.1",
     "pdf-parse": "^1.1.1",
