+++ packages/frontend/src/pages/Dashboard/AdvancedView.tsx
@@ -1,301 +1,116 @@
-import React, { useState, useRef } from 'react';
-import { Send, FileDown, Clock, MessageSquare } from 'lucide-react';
-import { jsPDF } from "jspdf";
+import { useState, useEffect, useContext } from "react";
+import DynamicChart, { ChartJsonData } from "./dynamicChart";
+import html2canvas from "html2canvas";
+import jsPDF from "jspdf";
+import { ChartContext } from "../../components/RouterRoot";
 
-interface ImageResponseCard {
-  title: string;
-  subtitle?: string;
-  buttons: Array<{
-    text: string;
-    value: string;
-  }>;
-}
+export const ChartHistory = () => {
+  const [charts, setCharts] = useState<ChartJsonData[]>([]);
+  const { chartJson } = useContext(ChartContext);
 
-interface Message {
-  type: 'user' | 'bot' | 'loading' | 'card';
-  content: string | ImageResponseCard;
-  timestamp: string;
-}
+  useEffect(() => {
+    // Load initial charts from sessionStorage
+    const loadCharts = () => {
+      try {
+        const storedCharts = sessionStorage.getItem("chartHistory");
+        if (storedCharts) {
+          setCharts(JSON.parse(storedCharts));
+        }
+      } catch (error) {
+        console.error("Error loading charts:", error);
+      }
+    };
 
-const AdvancedView: React.FC = () => {
-  const [messages, setMessages] = useState<Message[]>([]);
-  const [sessionId, setSessionId] = useState<string>("");
-  const [inputValue, setInputValue] = useState<string>("");
-  const [isLoading, setIsLoading] = useState<boolean>(false);
-  const chatRef = useRef<HTMLDivElement>(null);
+    loadCharts();
+  }, []);
 
-  const startSession = async (): Promise<string> => {
-    if (sessionId) return sessionId;
-    try {
-      const response = await fetch(`${import.meta.env.VITE_API_URL}/lex/start-session`, {
-        method: "POST",
+  // Update charts when chartJson changes
+  useEffect(() => {
+    if (chartJson && chartJson.length > 0) {
+      const latestChart = chartJson[chartJson.length - 1];
+
+      // Update state with new chart
+      setCharts(prevCharts => {
+        const newCharts = [latestChart, ...prevCharts]; // Newest chart at the top
+
+        // Update sessionStorage
+        try {
+          sessionStorage.setItem("chartHistory", JSON.stringify(newCharts));
+        } catch (error) {
+          console.error("Error saving to sessionStorage:", error);
+        }
+
+        return newCharts;
       });
-      const data = await response.json();
-      if (!data.sessionId) throw new Error("Session not started");
-      setSessionId(data.sessionId);
-      return data.sessionId;
-    } catch (error) {
-      console.error("Could not start session.", error);
-      throw error;
     }
-  };
+  }, [chartJson]);
 
-  const scrollToBottom = () => {
-    if (chatRef.current) {
-      setTimeout(() => {
-        chatRef.current?.scrollTo({
-          top: chatRef.current.scrollHeight,
-          behavior: 'smooth'
+  const exportHistoryAsPDF = async () => {
+    const historyEl = document.getElementById("chart-history");
+    if (historyEl && charts.length > 0) {
+      try {
+        const pdf = new jsPDF({
+          orientation: "portrait",
+          unit: "mm",
+          format: "a4",
         });
-      }, 100);
-    }
-  };
 
-  const handleSendMessage = async (message: string): Promise<void> => {
-    if (!message.trim()) return;
+        const pageWidth = pdf.internal.pageSize.getWidth();
+        const pageHeight = pdf.internal.pageSize.getHeight();
+        let currentHeight = 0;
 
-    setIsLoading(true);
-    try {
-      // Add user message
-      setMessages(prev => [...prev, {
-        type: 'user',
-        content: message,
-        timestamp: new Date().toLocaleString()
-      }]);
-      scrollToBottom();
+        for (const chartEl of Array.from(historyEl.children)) {
+          const canvas = await html2canvas(chartEl as HTMLElement, { 
+            scale: 2,
+            logging: false,
+            useCORS: true,
+          } as any);
+          
+          const imgData = canvas.toDataURL("image/png");
+          const imgProps = pdf.getImageProperties(imgData);
+          const imgHeight = (imgProps.height * pageWidth) / imgProps.width;
 
-      // Add loading message
-      setMessages(prev => [...prev, {
-        type: 'loading',
-        content: 'Thinking...',
-        timestamp: new Date().toLocaleString()
-      }]);
-      scrollToBottom();
+          // Check if the current chart fits on the current page
+          if (currentHeight + imgHeight > pageHeight) {
+            pdf.addPage(); // Start a new page if it doesn't fit
+            currentHeight = 0; // Reset current height for the new page
+          }
 
-      const sid = await startSession();
-      const lexResponse = await fetch(`${import.meta.env.VITE_API_URL}/lex/message-lex`, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({ message, sessionId: sid }),
-      });
-
-      if (!lexResponse.ok) throw Error("Could not send request.");
-
-      const result = await lexResponse.json();
-      const body = result.response;
-      console.log("lex response:", body);
-
-      // Replace loading message with actual response
-      if (body.messages && body.messages[0]) {
-        if (body.messages[0].contentType === "ImageResponseCard") {
-          setMessages(prev => prev.slice(0, -1).concat({
-            type: 'card',
-            content: body.messages[0].imageResponseCard,
-            timestamp: new Date().toLocaleString()
-          }));
-        } else {
-          if (!body.messages[0].content) throw Error("Empty message");
-          setMessages(prev => prev.slice(0, -1).concat({
-            type: 'bot',
-            content: body.messages[0].content,
-            timestamp: new Date().toLocaleString()
-          }));
-        }
-        scrollToBottom();
-      }
-
-    } catch (error) {
-      console.error("Error:", error);
-      setMessages(prev => {
-        const withoutLoading = prev.filter(msg => msg.type !== 'loading');
-        return [...withoutLoading, {
-          type: 'bot',
-          content: "An error occurred. Please try again.",
-          timestamp: new Date().toLocaleString()
-        }];
-      });
-      scrollToBottom();
-    } finally {
-      setIsLoading(false);
-      setInputValue("");
-    }
-  };
-
-  const shouldIncludeInReport = (content: string | ImageResponseCard): boolean => {
-    if (typeof content !== 'string') return false;
-    return !content.includes('?');
-  };
-
-  const exportToPDF = (): void => {
-    const doc = new jsPDF();
-    let yOffset = 20;
-
-    doc.setFontSize(16);
-    doc.text("Analysis Report", 20, yOffset);
-    yOffset += 20;
-
-    doc.setFontSize(12);
-    messages.forEach((msg) => {
-      if (msg.type === 'user' || (msg.type === 'bot' && shouldIncludeInReport(msg.content))) {
-        if (yOffset > 270) {
-          doc.addPage();
-          yOffset = 20;
+          pdf.addImage(imgData, "PNG", 0, currentHeight, pageWidth, imgHeight);
+          currentHeight += imgHeight + 10; // Update height for the next chart
         }
 
-        const content = typeof msg.content === 'string' ? msg.content : '';
-        const lines = doc.splitTextToSize(content, 170);
-
-        doc.text(lines, 20, yOffset);
-        yOffset += 10 * lines.length + 5;
+        pdf.save("chart-history.pdf");
+      } catch (error) {
+        console.error("Error exporting PDF:", error);
       }
-    });
-
-    doc.save("analysis-report.pdf");
-  };
-
-  const exportToWord = (): void => {
-    let content = '<html><body>';
-    content += '<h1>Analysis Report</h1>';
-
-    messages.forEach((msg) => {
-      if (msg.type === 'user' || (msg.type === 'bot' && shouldIncludeInReport(msg.content))) {
-        const prefix = msg.type === 'user' ? '<strong>User:</strong> ' : '<strong>Response:</strong> ';
-        const text = typeof msg.content === 'string' ? msg.content : '';
-        content += `<p>${prefix}${text}</p>`;
-      }
-    });
-
-    content += '</body></html>';
-
-    const blob = new Blob([content], { type: 'application/msword' });
-    const link = document.createElement('a');
-    link.href = URL.createObjectURL(blob);
-    link.download = 'analysis-report.doc';
-    link.click();
-    URL.revokeObjectURL(link.href);
-  };
-
-  const renderMessage = (msg: Message) => {
-    if (msg.type === 'loading') {
-      return (
-          <div className="flex justify-center">
-            <div className="animate-pulse text-gray-400">
-              {typeof msg.content === 'string' ? msg.content : ''}
-            </div>
-          </div>
-      );
     }
-
-    if (msg.type === 'card') {
-      const card = msg.content as ImageResponseCard;
-      return (
-          <div className="flex justify-start">
-            <div className="bg-primary rounded-lg p-4 max-w-[80%] border border-gray-700">
-              <h3 className="font-semibold text-lg mb-2">{card.title}</h3>
-              {card.subtitle && <p className="text-gray-300 mb-3">{card.subtitle}</p>}
-              <div className="flex flex-wrap gap-2">
-                {card.buttons.map((button, i) => (
-                    <button
-                        key={i}
-                        onClick={() => handleSendMessage(button.value)}
-                        className="bg-meta-3 text-white px-3 py-1.5 rounded hover:bg-meta-4 transition-colors"
-                    >
-                      {button.text}
-                    </button>
-                ))}
-              </div>
-              <p className="text-xs text-gray-400 mt-3 flex items-center gap-1">
-                <Clock className="w-3 h-3" />
-                {msg.timestamp}
-              </p>
-            </div>
-          </div>
-      );
-    }
-
-    const isUser = msg.type === 'user';
-    return (
-        <div className={`flex ${isUser ? 'justify-end' : 'justify-start'}`}>
-          <div className={`${isUser ? 'bg-meta-6' : 'bg-primary'} rounded-lg p-4 max-w-[80%] ${!isUser && 'border border-gray-700'}`}>
-            <p className="text-white">{typeof msg.content === 'string' ? msg.content : ''}</p>
-            <p className="text-xs text-gray-400 mt-2 flex items-center gap-1">
-              <Clock className="w-3 h-3" />
-              {msg.timestamp}
-            </p>
-          </div>
-        </div>
-    );
   };
 
   return (
-      <div className="flex flex-col h-[calc(100vh-160px)]">
-        <div className="bg-boxdark border-b border-gray-700 p-4">
-          <div className="max-w-4xl mx-auto flex justify-between items-center">
-            <h2 className="text-xl font-semibold flex items-center gap-2">
-              <MessageSquare className="w-5 h-5" />
-              Advanced Analysis
-            </h2>
-            <div className="flex gap-2">
-              <button
-                  onClick={exportToPDF}
-                  className="flex items-center gap-1 px-3 py-1.5 rounded-lg bg-meta-3 text-white hover:bg-meta-4 transition-colors"
-              >
-                <FileDown className="w-4 h-4" />
-                Export PDF
-              </button>
-              <button
-                  onClick={exportToWord}
-                  className="flex items-center gap-1 px-3 py-1.5 rounded-lg bg-meta-5 text-white hover:bg-meta-6 transition-colors"
-              >
-                <FileDown className="w-4 h-4" />
-                Export Word
-              </button>
-            </div>
-          </div>
-        </div>
-
-        <div className="flex-1 overflow-hidden">
-          <div className="max-w-4xl mx-auto h-full flex flex-col">
-            <div
-                ref={chatRef}
-                className="flex-1 overflow-y-auto p-4 space-y-6"
+    <div className="p-4">
+      {charts.length > 0 ? (
+        <>
+          <button 
+            onClick={exportHistoryAsPDF}
+            className="mb-4 px-4 py-2 bg-blue-500 text-white rounded hover:bg-blue-600"
+          >
+            Export History
+          </button>
+          <div id="chart-history" className="space-y-4">
+          {charts.map((chart, idx) => (
+            <div 
+              key={`chart-${idx}-${Date.now()}`} // Generate a unique key on the fly
+              className="p-4 border border-gray-200 rounded shadow-sm"
             >
-              {messages.map((msg, idx) => (
-                  <div key={idx}>
-                    {renderMessage(msg)}
-                  </div>
-              ))}
+              <DynamicChart jsonData={chart} />
             </div>
-
-            <div className="p-4 border-t border-gray-700">
-              <div className="flex gap-2 max-w-4xl mx-auto">
-                <input
-                    type="text"
-                    value={inputValue}
-                    onChange={(e) => setInputValue(e.target.value)}
-                    className="flex-1 p-3 rounded-lg bg-primary border border-gray-700 text-bodydark1 focus:outline-none focus:border-meta-3"
-                    placeholder="Type your message..."
-                    onKeyPress={(e) => {
-                      if (e.key === 'Enter' && !isLoading) {
-                        handleSendMessage(inputValue);
-                      }
-                    }}
-                />
-                <button
-                    onClick={() => handleSendMessage(inputValue)}
-                    disabled={isLoading}
-                    className="p-3 rounded-lg bg-meta-3 text-white hover:bg-meta-4 transition-colors disabled:opacity-50"
-                >
-                  <Send className="w-5 h-5" />
-                </button>
-              </div>
-            </div>
+          ))}
           </div>
-        </div>
-      </div>
+        </>
+      ) : (
+        <p className="text-gray-500">No chart history available</p>
+      )}
+    </div>
   );
 };
-
-export default AdvancedView;
